Добавил Свич на 2 и три колонки

+++ app/src/main/java/com/example/testappforppr/MainViewModel.kt
@@ -11,7 +11,7 @@ import kotlinx.coroutines.flow.Flow
 import java.math.BigInteger
 
 class MainViewModel (numbersRepositoryImpl : NumbersRepositoryImpl) : ViewModel() {
-    val primeNumbers: Flow<PagingData<Int>> = Pager(PagingConfig(pageSize = 20)){
+    val primeNumbers: Flow<PagingData<BigInteger>> = Pager(PagingConfig(pageSize = 20)){
         PrimeNumbersSource(numbersRepositoryImpl)
     }.flow
     val fibonacciNumbers: Flow<PagingData<BigInteger>> = Pager(PagingConfig(pageSize = 20)){
