+++ app/src/main/java/com/example/testappforppr/model/NumbersRepository.kt
@@ -3,6 +3,6 @@ package com.example.testappforppr.model
 import java.math.BigInteger
 
 interface NumbersRepository {
-    suspend fun getPrimeNumbers(page: Int) : List<Int>
+    suspend fun getPrimeNumbers(page: Int) : List<BigInteger>
     suspend fun getFibonacciNumbers(page: Int) :List<BigInteger>
 }
